Create Excel_Challenge_779 - Linked_in_Block_Shift.py
Shift 1: Starting with A, B, C, D, E, moving one space to the right pushes E out of the block. Move E to the front.
Shift 4: Starting with J, W, U, K, F, moving four spaces to the right pushes W, U, K, and F out of the block. Move W, U, K, and F to the front.
After completing the row shifts of 1, 4, and 3, execute the column shifts of 2, 3, 1, 5, and 4 in the same manner.
This is called 1 Set. 
The displayed result is the outcome of executing 50,000 sets. You should execute it minimum number of times to achieve the same result.